Add autosizing everywhere
